Cleaned up wayland module
