+++ src/wayland/mod.rs
@@ -1,179 +1,17 @@
 pub mod gamma_control;
 
-use std::ptr;
 use wayland_server::Resource;
 use wayland_server::sys::WAYLAND_SERVER_HANDLE;
-use wayland_sys::server::*;
-use wayland_sys::common::*;
-use self::gamma_control::{GammaControl, GammaHandler, GammaManagerHandler};
-use rustwlc::handle::{wlc_handle_from_wl_output_resource, WlcOutput, WlcView};
-use rustwlc::render::{wlc_output_set_gamma, wlc_output_get_gamma_size};
 use rustwlc::wayland;
-
-
-use nix::libc::{uint32_t, uint16_t, c_uint};
-use std::os::raw::c_void;
-use std::ptr::null_mut;
-
-use std::sync::{Mutex};
-
-static SET_GAMMA_ERROR: &'static str = "The gamma ramps don't have the same size!";
-
-// TODO Is this the best way really? Not use anything from the generated file?
-// TODO Move this to gamma_control.rs
-unsafe extern "C" fn gamma_control_send_gamma_size(resource: *mut wl_resource,
-                                                   size: uint16_t) {
-    ffi_dispatch!(WAYLAND_SERVER_HANDLE,
-                  wl_resource_post_event,
-                  resource,
-                  0,
-                  size as c_uint);
-}
-
-#[repr(C)]
-pub struct GammaControlManagerInterface {
-    destroy: unsafe extern "C" fn (client: *mut wl_client,
-                                   resource: *mut wl_resource),
-    get_gamma_control: unsafe extern "C" fn (client: *mut wl_client,
-                                             resource: *mut wl_resource,
-                                             id: u32,
-                                             output: *mut wl_resource)
-}
-
-#[repr(C)]
-pub struct GammaControlInterface {
-    destroy: unsafe extern "C" fn (client: *mut wl_client,
-                                   resource: *mut wl_resource),
-    set_gamma: unsafe extern "C" fn (client: *mut wl_client,
-                                     resource: *mut wl_resource,
-                                     red: *mut wl_array,
-                                     green: *mut wl_array,
-                                     blue: *mut wl_array),
-    reset_gamma: unsafe extern "C" fn (client: *mut wl_client,
-                                       resource: *mut wl_resource)
-}
-
-unsafe extern "C" fn set_gamma(client: *mut wl_client,
-                               resource: *mut wl_resource,
-                               red: *mut wl_array,
-                               green: *mut wl_array,
-                               blue: *mut wl_array) {
-    error!("SETTTTINGGG gamma");
-    if (*red).size != (*green).size || (*red).size != (*blue).size {
-        ffi_dispatch!(
-            WAYLAND_SERVER_HANDLE,
-            wl_resource_post_error,
-            resource,
-            0, // Invalid gamma
-            SET_GAMMA_ERROR.as_bytes().as_ptr() as *const i8);
-        warn!("Color size error, can't continue");
-        return
-    }
-    let r = (*red).data as *mut u16;
-    let g = (*green).data as *mut u16;
-    let b = (*blue).data as *mut u16;
-    let user_data = ffi_dispatch!(
-        WAYLAND_SERVER_HANDLE,
-        wl_resource_get_user_data,
-        resource) as *const _;
-    let output = WlcOutput(wlc_handle_from_wl_output_resource(user_data));
-    // TODO Make this less stupid to check if it's a null index
-    if output.as_view().is_root() {
-        warn!("Output thing was wrong");
-        return;
-    }
-    debug!("Setting gamma for output {:?}", output);
-    warn!("Setting gamma for output {:?}", output);
-    error!("Setting gamma for output {:?}", output);
-    wlc_output_set_gamma(output.0, ((*red).size / 2) as u16, r, g, b)
-
-}
-unsafe extern "C" fn reset_gamma(client: *mut wl_client,
-                                 resource: *mut wl_resource) {
-    // Do nothing
-}
-
-unsafe extern "C" fn destroy(wl_client: *mut wl_client,
-                resource: *mut wl_resource) {
-    warn!("Destroying resource!");
-    ffi_dispatch!(
-        WAYLAND_SERVER_HANDLE,
-        wl_resource_destroy,
-        resource
-    );
-}
-
-unsafe extern "C" fn get_gamma_control(client: *mut wl_client,
-                                       resource: *mut wl_resource,
-                                       id: uint32_t,
-                                       output: *mut wl_resource) {
-    let manager_resource = ffi_dispatch!(
-        WAYLAND_SERVER_HANDLE,
-        wl_resource_create,
-        client,
-        GammaControl::interface_ptr(),
-        GammaControl::supported_version() as i32,
-        id);
-    // check that the output is something we control from wlc
-    warn!("In the setup or whatever");
-    let wlc_output = WlcOutput(wlc_handle_from_wl_output_resource(output as *const _));
-    // TODO Make this less stupid to check if it's a null index
-    if wlc_output.as_view().is_root() {
-        warn!("This is triggering, dis bad?");
-        return;
-    }
-    let mut gamma_control = GAMMA_CONTROL.try_lock().unwrap();
-    let gamma_control_ptr = &mut *gamma_control as *mut _ as *mut c_void;
-    ffi_dispatch!(
-        WAYLAND_SERVER_HANDLE,
-        wl_resource_set_implementation,
-        manager_resource,
-        gamma_control_ptr,
-        output as *mut c_void,
-        None
-    );
-    warn!("Sending gamma size to thingie");
-    gamma_control_send_gamma_size(manager_resource,
-                                  wlc_output_get_gamma_size(wlc_output.0))
-}
-
-
-impl GammaControlManagerInterface {
-    pub fn new() -> Self {
-        GammaControlManagerInterface {
-            destroy: destroy,
-            get_gamma_control: get_gamma_control
-        }
-    }
-}
-
-impl GammaControlInterface {
-    pub fn new() -> Self {
-        GammaControlInterface {
-            destroy: destroy,
-            set_gamma: set_gamma,
-            reset_gamma: reset_gamma
-        }
-    }
-}
-
-lazy_static!(
-    static ref GAMMA_CONTROL_MANAGER: Mutex<GammaControlManagerInterface> =
-        Mutex::new(GammaControlManagerInterface::new());
-    static ref GAMMA_CONTROL: Mutex<GammaControlInterface> =
-        Mutex::new(GammaControlInterface::new());
-);
+use std::ptr;
 
 /// Initializes the appropriate handlers for each wayland protocol
 /// that Way Cooler supports.
 pub fn init_wayland_protocols() {
-    error!("Initializing wayland protocols");
+    debug!("Initializing wayland protocols");
     let w_display = wayland::get_display();
-    //declare_handler!(GammaHandler,
-    //                 gamma_control::Handler,
-    //                 gamma_control::GammaControl);
     unsafe {
-        // Initalize the gamma control manager
+        debug!("Initializing gamma control manager");
         ffi_dispatch!(WAYLAND_SERVER_HANDLE,
                       wl_global_create,
                       w_display as *mut _,
@@ -183,6 +21,4 @@ pub fn init_wayland_protocols() {
                       gamma_control::bind
         );
     }
-    error!("Created global!");
-    // TODO Doesn't handler get dropped?
 }
